requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai>=1.6.1
langchain>=0.0.350
langchain-openai>=0.0.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pandas==2.1.1
numpy==1.24.3
gunicorn==21.2.0 